314 - go full-screen on player double click
